Create World Cup Scores Calculator
I want to write some code on Python to calculate the scores of the AES team WC 2018 Competition.
We divided the team agents on 6 groups, 5 with 3 members and 1 with only 2. Each group saved their guessed scores for the Group Stage phase of the World Cup on and excell file. A total of 46 matches.
The score cards is:
- 3 points if you guessed the exact result.
- 2 points if you guessed the goal difference.
- 1 point if you guessed tendency.
- 0 points otherwise.